+++ iceoryx_hoofs/test/moduletests/test_cxx_filesystem.cpp
@@ -113,7 +113,7 @@ TEST(filesystem_test, permsSatisfiesBinaryExclusiveOrAssignmentOperationCorrectl
     EXPECT_THAT(static_cast<permsBaseType_t>(sut ^= TEST_VALUE_RHS), Eq(sutBaseValue ^= BASE_VALUE_RHS));
 }
 
-TEST(filesystem_test, permsWhenEverythingIsSetTheOutputPrintsEverything)
+TEST(filesystem_test, streamOperatorPrintsCorrectlyWhenEverythingIsSet)
 {
     ::testing::Test::RecordProperty("TEST_ID", "2bb4931f-6ef9-4089-88a1-bf263a931559");
     Logger_Mock loggerMock;
@@ -128,7 +128,7 @@ TEST(filesystem_test, permsWhenEverythingIsSetTheOutputPrintsEverything)
                    "special bits: {set_uid, set_git, sticky_bit}"));
 }
 
-TEST(filesystem_test, permsWhenNothingIsSetEveryEntryIsNone)
+TEST(filesystem_test, streamOperatorPrintsCorrectlyWhenNothingIsSet)
 {
     ::testing::Test::RecordProperty("TEST_ID", "2b50cb56-6dae-4514-bd77-791f81f6adca");
     Logger_Mock loggerMock;
@@ -142,7 +142,7 @@ TEST(filesystem_test, permsWhenNothingIsSetEveryEntryIsNone)
                 Eq("owner: {none},  group: {none},  others: {none},  special bits: {none}"));
 }
 
-TEST(filesystem_test, permsWhenSomeOrSetTheOutputIsCorrect)
+TEST(filesystem_test, streamOperatorPrintsCorrectlyWhenPartiallySet)
 {
     ::testing::Test::RecordProperty("TEST_ID", "94e647b7-242b-4fe3-bccd-2fde9e091e8e");
     Logger_Mock loggerMock;
@@ -156,4 +156,19 @@ TEST(filesystem_test, permsWhenSomeOrSetTheOutputIsCorrect)
                 Eq("owner: {write, execute},  group: {read, execute},  others: {read, write, execute},  special bits: "
                    "{sticky_bit}"));
 }
+
+TEST(filesystem_test, streamOperatorPrintsCorrectlyWhenSetToUnknown)
+{
+    ::testing::Test::RecordProperty("TEST_ID", "bcfd29e1-84d9-11ec-9e17-5405db3a3777");
+    Logger_Mock loggerMock;
+    {
+        auto logStream = iox::log::LogStream(loggerMock);
+        logStream << perms::unknown;
+    }
+
+    ASSERT_THAT(loggerMock.m_logs.size(), Eq(1U));
+    EXPECT_THAT(loggerMock.m_logs[0].message, Eq("unknown permissions"));
+}
+
+
 } // namespace